Customize the CONTRIBUTING.md
The architecture of coderadar is updated. Furthermore the database is
changed to the graph database Neo4j. Thus the CONTRIBUTING.md must be
adapted. In order to simplify the development of coderadar for everyone,
a Git Workflow is also defined. For following changes of the file the
file was adapted to the usual Markdown conventions (line break after
point, blank line after heading).