Customize the DeleteAnalyzerConfigurationUseCase
Since the UseCases are uniquely named, the variables do not have to have
long names that are redundant. The interface/class names can be used to
identify what they represent. A method has been added to the port that
allows an analyzer configuration to be deleted using the entity. When
deleting, the project ID is not required, since the ID of the
AnalyzerConfiguration is enough to uniquely identify it.