Update the query for finding all commits belonging to a project
The new query returns all commits. In the old query, two types of commit were not taken into account:

- Commits in which no files were changed.

- Commits in which only files that were in submodules were changed.

If a project consists only of commits in which no files have been changed, the commits are not collected. However, this happens very rarely and an analysis of such a project makes no sense. Adapting to this situation would involve the domain (a relationship between the project and the first commit would have to be added).