Adapted coderadar-core tests
Simplified tests by using the Mockito Extension for JUnit Jupiter so the
@Mock annotation can be used instead of initializing each mock by calling
the Mockito#mock() method. Added a lombok.config file to configure lombok
to return the object reference instead of void on generated setters to
simplify chaining of setter calls. Rewrote and modified most tests in coderadar
core and used AssertJ assertions.

To fix the issue of not knowing which workdir name for projects
will get generated in tests, a new inner funcational Interface called
WorkdirNameGenerator` was introduced to replace the implementation
in test runs.